Adds the ability to use parenting instead of joining when loading GLB.

+++ kubric/renderer/blender.py
@@ -36,6 +36,24 @@ import tensorflow as tf
 logger = logging.getLogger(__name__)
 
 
+def add_top_level_empty_parent(name: str = "Empty") -> bpy.types.Object:
+  """Adds an empty parent to scene and makes it the parent of all objects.
+
+  Args:
+    name: The name of the empty parent.
+
+  Returns:
+    The newly created empty parent.
+  """
+  parent_obj = bpy.data.objects.new(name, None)
+  parent_obj.rotation_mode = "QUATERNION"
+  bpy.context.scene.collection.objects.link(parent_obj)
+  for obj in bpy.context.scene.objects:
+    if obj != parent_obj and obj.parent is None:
+      obj.parent = parent_obj
+  return parent_obj
+
+
 # noinspection PyUnresolvedReferences
 class Blender(core.View):
   """ An implementation of a rendering backend in Blender/Cycles."""
@@ -422,29 +440,40 @@ class Blender(core.View):
                   location=True, rotation=True, scale=True
               )
 
-            # gltf files often contain "Empty" objects as placeholders for camera / lights etc.
-            # here we are interested only in the meshes, we filter these out and join all meshes into one.
-            mesh = [m for m in bpy.context.selected_objects if m.type == "MESH"]
-            assert mesh
-            for ob in mesh:
-              ob.select_set(state=True)
-              bpy.context.view_layer.objects.active = ob
+            if obj.use_parenting_instead_of_join:
+              parent_obj = add_top_level_empty_parent(obj.uid)
+              bpy.ops.object.select_all(action="DESELECT")
+              parent_obj.select_set(state=True)
+            else:
+              # Legacy loader which relies on JOIN. NOTE: This will destroy
+              # things like animations.
+              # gltf files often contain "Empty" objects as placeholders for
+              # camera / lights etc.
+              # here we are interested only in the meshes, we filter these out
+              # and join all meshes into one.
+              mesh = [
+                  m for m in bpy.context.selected_objects if m.type == "MESH"
+              ]
+              assert mesh
+              for ob in mesh:
+                ob.select_set(state=True)
+                bpy.context.view_layer.objects.active = ob
 
-            # make sure one of the objects is active, otherwise join() fails.
-            # see https://blender.stackexchange.com/questions/132266/joining-all-meshes-in-any-context-gets-error
-            bpy.context.view_layer.objects.active = mesh[0]
-            bpy.ops.object.join()
+              # make sure one of the objects is active, otherwise join() fails.
+              # see https://blender.stackexchange.com/questions/132266/joining-all-meshes-in-any-context-gets-error
+              bpy.context.view_layer.objects.active = mesh[0]
+              bpy.ops.object.join()
 
-            # Make sure to delete all remaining non-mesh objects. Note that for
-            # some reason deleting the non-mesh objets before joining removes
-            # parts of the meshes in some cases.
-            non_mesh_objects = [
-                obj
-                for obj in bpy.context.selected_objects
-                if obj.type != "MESH"
-            ]
-            with bpy.context.temp_override(selected_objects=non_mesh_objects):
-              bpy.ops.object.delete()
+              # Make sure to delete all remaining non-mesh objects. Note that
+              # for some reason deleting the non-mesh objets before joining
+              # removes parts of the meshes in some cases.
+              non_mesh_objects = [
+                  obj
+                  for obj in bpy.context.selected_objects
+                  if obj.type != "MESH"
+              ]
+              with bpy.context.temp_override(selected_objects=non_mesh_objects):
+                bpy.ops.object.delete()
 
             assert len(bpy.context.selected_objects) == 1
             blender_obj = bpy.context.selected_objects[0]
@@ -473,7 +502,8 @@ class Blender(core.View):
 
     # deactivate auto_smooth because for some reason it lead to no smoothing at all
     # TODO: make smoothing configurable
-    blender_obj.data.use_auto_smooth = False
+    if hasattr(blender_obj.data, "use_auto_smooth"):
+      blender_obj.data.use_auto_smooth = False
 
     register_object3d_setters(obj, blender_obj)
     obj.observe(AttributeSetter(blender_obj, "active_material",
